backend/pyproject.toml
[project]
name = "fault-watch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioredis>=2.0.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "redis>=5.0.1",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn[standard]>=0.34.3",
    "psycopg2-binary>=2.9.9",
    "fastapi-utils>=0.8.0",
    "typing-inspect>=0.9.0",
    "temporalio>=1.5.0",
    "python-slugify>=8.0.4",
]

[dependency-groups]
dev = [
    "alembic>=1.16.1",
]
